CTCP-4573: Gross weight page

+++ app/utils/transformers/GoodsMeasureTransformer.scala
@@ -18,7 +18,8 @@ package utils.transformers
 
 import generated.GoodsMeasureType03
 import models.{Index, UserAnswers}
-import pages.{GrossWeightPage, NetWeightPage}
+import pages.NetWeightPage
+import pages.houseConsignment.index.items.GrossWeightPage
 
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
@@ -29,7 +30,7 @@ class GoodsMeasureTransformer @Inject() (implicit ec: ExecutionContext) extends 
     goodsMeasure match {
       case Some(GoodsMeasureType03(grossMass, netMass)) =>
         val pipeline: UserAnswers => Future[UserAnswers] =
-          set(GrossWeightPage(hcIndex, itemIndex), grossMass.toDouble) andThen
+          set(GrossWeightPage(hcIndex, itemIndex), grossMass) andThen
             set(NetWeightPage(hcIndex, itemIndex), netMass.map(_.toDouble))
 
         pipeline(userAnswers)
